# backend/apiService/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /user/info:
        get:
            tags:
                - UserService
            description: 获取用户信息
            operationId: UserService_GetUserInfo
            parameters:
                - name: userId
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/apiService.svapi.GetUserInfoResponse'
        put:
            tags:
                - UserService
            description: 更新用户信息
            operationId: UserService_UpdateUserInfo
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/apiService.svapi.UpdateUserInfoRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/apiService.svapi.UpdateUserInfoResponse'
...
components:
    schemas:
...
        apiService.svapi.GetUserInfoResponse:
            type: object
            properties:
                user:
                    $ref: '#/components/schemas/apiService.svapi.User'
...
        apiService.svapi.UpdateUserInfoRequest:
            type: object
            properties:
                userId:
                    type: string
                    description: '@gotags: json:"user_id,omitempty,string"'
                name:
                    type: string
                avatar:
                    type: string
                backgroundImage:
                    type: string
                signature:
                    type: string
        apiService.svapi.UpdateUserInfoResponse:
            type: object
            properties: {}
        apiService.svapi.User:
            type: object
            properties:
                id:
                    type: string
                    description: '@gotags: json:"id,omitempty,string"'
                name:
                    type: string
                avatar:
                    type: string
                backgroundImage:
                    type: string
                signature:
                    type: string
                mobile:
                    type: string
                email:
                    type: string
                followCount:
                    type: string
                    description: '@gotags: json:"follow_count,omitempty,string"'
                followerCount:
                    type: string
                    description: '@gotags: json:"follower_count,omitempty,string"'
                totalFavorited:
                    type: string
                    description: '@gotags: json:"total_favorited,omitempty,string"'
                workCount:
                    type: string
                    description: '@gotags: json:"work_count,omitempty,string"'
                favoriteCount:
                    type: string
                    description: '@gotags: json:"favorite_count,omitempty,string"'
...